Fix dirty state calculation in inlineEditor.
The recent optimization for inlineEditor resulted in fields that could
not be saved, due to the dirty state not yet being updated.

Now, we force a calculation when we need the value. It will result in an
extra calculation, since a _.throttle can't be canceled, but it
shouldn't have any noticeable impact, as it occurs only on save/cancel
or when specifically requesting the dirty state.

Reviewed at https://reviews.reviewboard.org/r/6433/