Adding support for alphabetic pagination to the datagrid.
Updating the paginator in datagrid to allow alphabetic querying to be turned on
by providing the DataGrid object with the current_letter parameter (default
none).

If the current_letter is provided to the datagrid (usually by some child grid
calling the super initializer) the datagrid will simply load a second template
case added paginator.html.

It is the job of the child class (UserGrid in reviewboard) to process the URI,
perform filtering on the queryset, and handle errors.

Testing Done:
- Unit tests all pass
- All button cases manually tested. (Forward, back, first and last arrows,
  each letter).
- Verified that  changes to not affect any other module that inherits from
  datagrid (such as the groups list, or the all review requests list, the
  current 'master' of reviewboard with no changes to the userGrid).

Bugs closed: 829

Reviewed at https://reviews.reviewboard.org/r/6369/