Add sandboxing for DataGrid columns.
Extensions that create a Column subclass (using the DataGridColumnsHook) can
throw exceptions inside Djblets. The DataGridColumnsHook hook and the parts of
Djblets that call those methods have been sandboxed.

Now when a Column subclass throws an exception, Djblets logs the errors with
enough information to find which method in the Column subclass threw the
exception.

Testing Done:

Unit tests have been written to make sure that functions from a Column subclass
have been called, and when an exception is thrown it gets logged. Tests for
DataGridColumnsHook test that the Column subclass when an exception is thrown
while registering and unregistering a Column subclass it gets logged.

The tests fail without the sanboxing, and succeed with it.

Reviewed at https://reviews.reviewboard.org/r/6502/