Refactor the alphabetic paginator changes into new AlphanumericDatagrid subclass.
This changes the implementation of the alphabetic paginator to be a subclass of
the Datagrid (AlphanumericDatagrid). The paginator(s) are no longer rendered
using template tags, but instead a render_paginator function, which uses
settings that are overloaded by the alphanumericDataGrid.

Testing Done:
- Unit tests all pass
- All button cases manually tested. (Forward, back, first and last arrows, each
  letter).
- Verified that  changes to not affect any other module that inherits from
  datagrid (such as the groups list, or the all review requests list, the
  current 'master' of reviewboard with no changes to the userGrid).

Bugs closed: 829

Reviewed at https://reviews.reviewboard.org/r/6521/