Update to jQuery 1.11
This change updates djblets to use jQuery 1.11. This also brings in the
jQuery-migrate plugin, which provides some compatibility implementations for
features which were removed from core jQuery.

The biggest changes within remove some compatibility-shims for older browsers
(notably IE8, which we haven't supported in a while). Other changes involve
some changes to `.attr()` vs `.prop()`, and a test to make sure we don't try to
append to nonexistant parts of the DOM (which is only a problem when running
js-tests).

Testing done:
- Ran reviewboard js-tests.
- Grepped through the javascript code and audited against the list of
  backwards-incompatible changes in jQuery 1.9 (1.10 and 1.11 are pretty
  trivial in terms of incompatible changes).

Reviewed at https://reviews.reviewboard.org/r/6698/