Fix UnicodeDecodeError when log contains an umlaut (Ö,Ü,Ä)
If the log file contains umlauts (for example, in the case of some built-in
exceptions when running on a non-English locale), the djblets log viewer will
crash with a UnicodeDecodeError. The fix is to open it using the 'codecs'
module with a specified encoding.

Testing done:
The change was validated on Python 2.7.9 x64 Win

Fixes bug 3760.