New Error for Djblets WebApi - DUPLICATE_ITEM
There are times when users are creating anything that are meant to be unique. To name a few, they may include usernames, account names, or ids. As they are meant to be unique, they may run into issues where the particular item they want to create is already taken. A new error has been added to djblets webapi to account for this particular scenario.

Reviewed at https://reviews.reviewboard.org/r/6968/