Make 'thumbnail' tag more capable.
Our existing thumbnailer code was pretty limited, and only allowed people to
pass in a certain image size as a string. This change improves it, especially
for the case where it's being used from Python code. Callers can now pass in
the thumbnail size as a 2-tuple of width and height. If the height is omitted,
it will be calculated from the width to preserve the aspect ratio of the image.

Testing done:
Used this in conjunction with my big file attachment redesign.

Reviewed at https://reviews.reviewboard.org/r/7114/