Add support in datagrids to show content when the grid is empty.
Custom datagrid listview.html templates can now override a
`datagrid_empty` block to display custom content when the datagrid is
empty. This is useful for displaying a message instead of an empty
datagrid, possibly with instructions on creating something.

The cell is wrapped in a row and cell with classes that can be styled
specially. These won't show the hover effect, like other datagrid rows.

By default, nothing is shown, for backwards-compatibility.

Testing Done:
Made use of this in a change for Review Board. Saw my message when the
datagrid was empty, but not when populated.

Reviewed at https://reviews.reviewboard.org/r/7046/