Add a utility function for encoding ETags to SHAs.
This adds an `encode_etag()` function that takes a string and turns it
into a SHA. It's meant for use when setting ETag headers.

The `WebAPIResource` code now uses this as well.

Testing Done:
Unit tests pass.

Used this in an upcoming change.

Reviewed at https://reviews.reviewboard.org/r/7168/