Add better flexibility to the datagrid templates.
The datagrid templates now offer a couple of new blocks for controlling
the styling of the page.

One is the datagrid_titlebox block, which contains the matching div.
This can be used to provide additional content before, after, or in
place of the titlebox without needing to resort to hacks.

The other is the "paginator" block, which wraps the paginator. This
allows the paginator to be removed or replaced.

Testing Done:
Made use of these in another change.

Reviewed at https://reviews.reviewboard.org/r/7215/