Improve positioning and scrolling of datagrid menus.
The datagrid menu placement was pretty imprecise, and didn't work very
well when auto-fitting the datagrid. The menu would go off the page and
couldn't be scrolled, and it also just didn't line up very well with the
right-hand side of the datagrid on all browsers. Along with all that, it
was pretty crowded and ugly.

A lot of the problems were due to a less-than-optimal calculation
combined with fractional pixels for the containers.

We now rely on CSS positioning to dock the menu on the right-hand side.
To work around any borders in the datagrid, the menu has been moved into
the datagrid-main div.

The height is set when we auto-fit, ensuring that it doesn't grow beyond
the confines of the table body.

The CSS rules have been updated to allow for scrolling within the menu,
and to give enough breathing room for the scrollbar (to avoid any issues
with a vertical scrollbar overlapping content, triggering a horizontal
scrollbar).

Some ugliness has also been cleaned up in the display. Namely, there's a
few more pixels of breathing room between entries, and they no longer
wrap (allowing the menu to widen to fit the contents appropriately).

Testing Done:
Tested the menu and all previous off-by-one issues we had in Chrome,
Firefox, and Safari on Retina and non-Retina displays.

Tested the Dashboard, All Review Requests, and Users pages in Review Board.

Reviewed at https://reviews.reviewboard.org/r/7213/