Exclude gridonly param when rendering paginator.
This is a fix for issue 3794 where editing the columns in a datagrid would lead to page rendering problems when switching to a different page via the paginator.

The issue is the inclusion of the 'gridonly=1' parameter in the paginator URLs that result after editing the columns.  `render_paginator()` is updated to exclude the `gridonly` parameter when producing the new URL query string.

(Note: I'm a little out of my comfort zone here, but this seems to be a sufficient solution.)

Testing Done:
Tested with internal dev server and confirmed that switching pages via the paginator no longer results in rendering issues.

Bugs closed: 3794

Reviewed at https://reviews.reviewboard.org/r/7230/