Add a new querystring_with tag to help with building links.
We often have cases where we want to create a link that will go back to the
same page, but with a single different querystring parameter. Traditionally
these links have been hard-coded, and doing so will drop all the other things
in the querystring. We do have a helper utility for doing what we want in
python (get_url_params_except), which is used in some places.

This change adds a new templatetag, `querystring_with`, that will build a new
querystring that has everything in the current one, but with the value of a
single variable changed.

Testing done:
Used this with other changes.

Reviewed at https://reviews.reviewboard.org/r/7268/