Fix loading large data with missing chunks from the cache.
The cache support was recently overhauled to support generators, but
this caused the `MissingChunkError` to be propagated to the code iterating
through results, rather than the code initially fetching the data from
cache.

To fix this, I've split the function into two: One that does the initial
fetching and missing chunk detection, and one that iterates through
results.

A unit test was added that failed with a propagated `MissingChunkError`
before this fix, and succeeded in falling through to the cache data
generation function after the fix.

Testing Done:
The new unit test failed before, and passed after. Also tested manually.

Reviewed at https://reviews.reviewboard.org/r/7272/