Restore CMD-click/shift-click/etc. behavior for datagrid rows.
The recent work to make entire rows clickable in the datagrid preventing
browser shortcuts for opening in new tabs/windows from working. This was
because we used an onclick handler, instead of letting the browser do
all the work.

We now wrap each cell in an <a> instead. If the cell already defined a
URL, that takes precedence, but it otherwise defaults to the URL of the
object that row represents.

Testing Done:
Tested keyboard shortcuts and right-click menu on the links. Saw it do the
right thing in all cases.

Reviewed at https://reviews.reviewboard.org/r/7283/