Redirect to a visible config page if the current one doesn't exist.
On a config forms page, if the selected page doesn't exist (for example, if it
decides it should hide itself), the view would end up showing the nav and
nothing else. This is kind of a crummy experience, especially in the case where
the user was on one page but then it decided to hide itself.

This change adds a check to make sure we're trying to navigate to a visible
page, and if not, will select the first one in the list.

Testing done:
Used this with a config page that would hide itself when requested. Instead of
saving and then showing an empty page, it saves and then shows the first page
in the nav.

Reviewed at https://reviews.reviewboard.org/r/7355/