Add a template filter for splitting a string.
Django lacks any template filters for splitting a string, which is a
handy thing to do sometimes. This adds a 'split' filter that can split
by a delimiter (defaulting to a comma), returning a list to iterate on.

Testing Done:
Used this in another change.

Reviewed at https://reviews.reviewboard.org/r/7353/