Support fractional values in $.getExtents().
`$.getExtents()` was normalizing everything to an integer, which would
affect layouts when dealing with browsers rendering things using
fractions of pixels. We now parse the value as a float if it's not a
integer.

Testing Done:
Made use of this in some other changes in Review Board. It solved the
problems I was having.

Been running this for a week, and I haven't seen any negative side-effects.

Reviewed at https://reviews.reviewboard.org/r/7390/