Fix up the base Djblets testing support for Django 1.8.
This gets the test harness in shape to test against Django 1.8. The test
runner has some additional setup it's now responsible for, and there's
middleware we don't need that doesn't exist in Django these days.

That's the easy part.

The mixin for loading test-specific models needed to be updated to work
with the new app/model registries in Django 1.7+. These are a much nicer
API for the sorts of things we had to hack around before, but isn't
really documented for the purposes we're dealing with.

We essentially now have to copy over all the models into a virtual
'models' module (which was already being set up), and then use the new
API for registering the models, pushing a new set of installed apps, and
updating the database using the 'migrate' management command.

The result is that the basic test harness works again, in both Django
1.6 and 1.8. Future changes will fix the individual tests.

Testing Done:
Along with the other upcoming changes, all unit tests pass on Django 1.6
and 1.8.

Reviewed at https://reviews.reviewboard.org/r/7418/