Add support in datagrids for simpler QuerySet-like objects.
Datagrids are now a bit less QuerySet-specific. They still take
QuerySet-like objects, but it's no longer necessary for them to support
`distinct()`, `values_list()`, or to have a `model` attribute.

This is primarily intended for use with Haystack's `SearchQuerySet`, which
can now be used to back the results of a datagrid. In this case, a
search result will be paginated, with each page turning into a typical
model query based on that page's IDs.

Testing Done:
Tested with standard datagrids with a real Queryset, and with a SearchQuerySet.

Reviewed at https://reviews.reviewboard.org/r/7436/