Skip documenting members listed in __deprecated__.
Sphinx will try to skip any members coming in through an import, but if
they're listed in `__all__`, it's assumed they should be documented. We
don't want this for any modules that are providing compatibility
imports, as it's just confusing.

Our excluder extension now looks for a `__deprecated__` member of the
module, and will exclude anything found in that list. This makes it easy
for us to choose what in a module we want to exclude and what we don't.

Testing Done:
Used this to exclude some modules I'm deprecating.

Reviewed at https://reviews.reviewboard.org/r/7470/