Split djblets.webapi.auth into multiple modules.
I have some new code coming soon for djblets.webapi, and it's going to
result in some of these files being even more crowded. To prepare for
that, I'm splitting up a few of the modules into more fine-grained,
purpose-specific modules. This is the first pass.

The auth backend classes now live in sub-modules of
djblets.webapi.auth.backends. The old login/logout views are in
djblets.webapi.auth.views.

The old djblets.webapi.auth will continue to work, but will emit a
warning when imported.

Along with this, I've updated the documentation for the new API docs, so
that we have something useful for consumers/contributors of this code.

Testing Done:
Unit tests for both Djblets and Review Board pass.

Built the docs and looked through all of them.

Reviewed at https://reviews.reviewboard.org/r/7471/