Split djblets.webapi.resources into sub-modules.
This is the second part of the djblets.webapi cleanup. This change takes
djblets.webapi.resources, which contained WebAPIResource, UserResource,
GroupResource, RootResource, and all the resource registration
functions, and moves them out into new submodules.

This will make it easier to bundle and maintain new built-in resources
for different purposes, and to keep the code more organized.

The old djblets.webapi.resources still works. It imports all the
original classes/functions, and triggers a deprecation warning when
imported.

Docstrings have been updated for most of this to conform with the new
codebase docs format we're using, but I didn't do this everywhere.
WebAPIResource is largely untouched, except for formatting errors for
lists.

Testing Done:
Unit tests pass on both Djblets and Review Board.

Built the docs and looked through them. Not everything is perfect, but
all the new/fixed content looks correct.

Reviewed at https://reviews.reviewboard.org/r/7472/