Add a foundation to help with writing API test suites.
This ports over some of Review Board's common API testing functionality
into Djblets, to help with writing fully-featured test suites. It
provides a mixin for test cases that provides functions for making API
calls and verifying the results, and for helping with docstring
generation.

Not all of Review Board's handy infrastructure made it over here, due to
how specific much of it is to Review Board, but with this change, it's
easy to begin writing new test suites.

Testing Done:
Ported Review Board's test suite over to this, and all unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/7489/