Add a class for handling the registration and lookup of API resources.
In Review Board, we had a nice little class that would handle the
loading of resources on-demand, and would allow for resources to
reference each other sanely without worrying about import loops.

This change brings the base of that class into Djblets, so that any
consumer of the API can use it. It's now easy to get this same
functionality in all applications interfacing with djblets.webapi.

Testing Done:
Made use of this in Review Board, and all unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/7477/