Remove warning messages from importing auth or resources
The `DeprecationWarning` warnings were intended to inform that the old
`djblets.webapi.auth` and `djblets.webapi.resources` modules are
deprecated and all functionality is stored in submodules. However,
these messages are always printed, even though the imports are updated.
This patch removes these deprecation warnings as they will always be
printed and are not helpful.

This patch also fixes one import that was not updated.

Testing Done:
No deprecation warnings are printed with this patch applied.

Reviewed at https://reviews.reviewboard.org/r/7516/